Do nor trim reads if trim_reads is false
It might be posible to use checkpoints and aggregate rule but it is eaiser to just add the flag and force the rule to be local when created the symlinks

+++ workflow/rules/detect_adapters.smk
@@ -8,26 +8,35 @@ rule detect_adapters:
         "logs/detect_adapters_{sample}.log"
     container:
         config["atria_image"]
+    localrule: 
+        not config["trim_reads"]
+    params:
+        do_trimming = config["trim_reads"]
     threads: 8
     shell:
         """
         mkdir -p $(dirname {output.adapter_txt})
+        if [[ params.do_trimming == "true" ]]; then
 
-        # Run Atria to detect adapters.
-        atria -r {input.fq1} -R {input.fq2} \
-            -o $(dirname {output.adapter_txt})/temp_{wildcards.sample} --detect-adapter --threads {threads} \
-            > {log} 2>&1
+            # Run Atria to detect adapters.
+            atria -r {input.fq1} -R {input.fq2} \
+                -o $(dirname {output.adapter_txt})/temp_{wildcards.sample} --detect-adapter --threads {threads} \
+                > {log} 2>&1
 
-        # Locate the detected adapter summary file
-        temp_file=$(find $(dirname {output.adapter_txt})/temp_{wildcards.sample}/ -type f -name "*.txt" | head -n 1)
+            # Locate the detected adapter summary file
+            temp_file=$(find $(dirname {output.adapter_txt})/temp_{wildcards.sample}/ -type f -name "*.txt" | head -n 1)
 
-        # Move the detected adapter file or create an empty one
-        if [[ -n "$temp_file" && -f "$temp_file" ]]; then
-            mv "$temp_file" {output.adapter_txt}
+            # Move the detected adapter file or create an empty one
+            if [[ -n "$temp_file" && -f "$temp_file" ]]; then
+                mv "$temp_file" {output.adapter_txt}
+            else
+                echo "Warning: No adapters detected for sample {wildcards.sample}. Creating an empty file." >> {log}
+                touch {output.adapter_txt}
+            fi
+            # Remove the temporary directory
+            rm -rf $(dirname {output.adapter_txt})/temp_{wildcards.sample}
         else
-            echo "Warning: No adapters detected for sample {wildcards.sample}. Creating an empty file." >> {log}
+            echo "Skipping as 'trim_reads' is false. Creating an empty file." >> {log}
             touch {output.adapter_txt}
         fi
-        # Remove the temporary directory
-        rm -rf $(dirname {output.adapter_txt})/temp_{wildcards.sample}
         """
